pyproject.toml
[tool.poetry]
name = "simbirsoft"
version = "0.1.0"
description = ""
authors = ["Roman Korzhenko <savroful@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.14.0"
pytest = "^7.4.2"
allure-pytest = "^2.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
